# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is Documentation from the project assignment list application",
      "version": "1.0.0",
      "title": "Assignment List",
      "contact": {
        "email": "luis.gustavo.282000@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
      "/categories": {  
        "get": {
          "tags": [
            "Categories"
          ],
          "summary": "Find Categories",
          "description": "Get All Categories",
          "operationId": "categories",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Category"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        },
        "post": {
          "tags": [
            "Categories"
          ],
          "summary": "Add new Category",
          "description": "",
          "operationId": "addCategory",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Category used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Category"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/categories/{id}": {
        "get": {
          "tags": [
            "Categories"
          ],
          "summary": "Find Categories by ID",
          "description": "Get Category By ID",
          "operationId": "getCategoryById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Category",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Category"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        },
        "put": {
          "tags": [
            "Categories"
          ],
          "summary": "Update Category",
          "description": "",
          "operationId": "updateCategory",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Category",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Category used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Category"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        },
        "delete": {
          "tags": [
            "Categories"
          ],
          "summary": "Delete Categories by ID",
          "description": "Delete Category By ID",
          "operationId": "deleteCategoryById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Category",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Category"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        }
      },
...
      "/tasks": {
        "get": {
          "tags": [
            "Tasks"
          ],
          "summary": "Find Tasks",
          "description": "Get All Tasks",
          "operationId": "tasks",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Task"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        },
        "post": {
          "tags": [
            "Tasks"
          ],
          "summary": "Add new Tasks",
          "description": "",
          "operationId": "addTasks",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Tasks used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Task"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/tasks/{id}": {
        "get": {
          "tags": [
            "Tasks"
          ],
          "summary": "Find Tasks by ID",
          "description": "Get Tasks By ID",
          "operationId": "getTaskById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Tasks",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Task"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        }
      },
      "/tasks/approved/{id}": {
        "post": {
          "tags": [
            "Tasks"
          ],
          "summary": "Approved Task",
          "description": "Approved Tasks",
          "operationId": "tasks approved",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Task",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Task"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        }
      },
      "/tasks/reject/{id}": {
        "post": {
          "tags": [
            "Tasks"
          ],
          "summary": "Reject Task",
          "description": "Reject Tasks",
          "operationId": "tasks reject",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Task",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Task"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        }
      }
    },
    "definitions": {
      "Category": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "xml": {
          "name": "Category"
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "client": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }